+++ src/main/java/com/example/bakery/repositories/OrdersRepository.java
@@ -8,15 +8,15 @@ import java.util.List;
 import java.time.LocalDateTime;
 
 @Repository
-public interface OrdersRepository extends JpaRepository<Orders, Integer> {
+public interface OrdersRepository extends JpaRepository<Orders, Long> {
     // Find orders by user ID
-    List<Orders> findByUserId(Integer userId);
+    List<Orders> findByUserId(Long userId);
 
     // Find orders created after a certain date
     List<Orders> findByCreatedAtAfter(LocalDateTime date);
 
     // Find orders by user ID and created after a certain date
-    List<Orders> findByUserIdAndCreatedAtAfter(Integer userId, LocalDateTime date);
+    List<Orders> findByUserIdAndCreatedAtAfter(Long userId, LocalDateTime date);
 
     // You can add more custom query methods as needed
 }
