+++ src/main/java/com/example/bakery/repositories/ReceiverRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReceiverRepository extends JpaRepository<Receivers, Integer> {
+public interface ReceiverRepository extends JpaRepository<Receivers, Long> {
     // You can add custom query methods here if needed
 }
