+++ src/main/java/com/example/bakery/services/UsersService.java
@@ -2,8 +2,8 @@ package com.example.bakery.services;
 
 import com.example.bakery.dto.RegisterRequest;
 import com.example.bakery.models.User;
+import com.example.bakery.repositories.UserRepository;
 import com.example.bakery.models.Role;
-import com.example.bakery.repository.UserRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -34,7 +34,7 @@ public class UsersService {
         return userRepository.save(newUser);
     }
 
-    public User getUserById(Integer id) {
+    public User getUserById(Long id) {
         return userRepository.findById(id).orElse(null);
     }
 }
